Move objects to individual files
