+++ MpesaDaraja/Models/AccessToken.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MpesaDaraja.Models
+{
+    public class AccessToken
+    {
+        public string access_token { get; set; }
+        public string expires_in { get; set; }
+    }
+}
