+++ MpesaDaraja/Models/DarajaResponse.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MpesaDaraja.Models
+{
+    public class DarajaResponse
+    {
+        public string OriginatorConverstionID { get; set; }
+        public string ConversationID { get; set; }
+        public string ResponseDescription { get; set; }
+    }
+}
