+++ MpesaDaraja/Models/RootObj.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MpesaDaraja.Models
+{
+    public class RootObj
+    {
+        public string ConversationID { get; set; }
+        public string OriginatorConversationID { get; set; }
+        public string OriginatorCoversationID { get; set; }
+        public string OriginatorConverstionID { get; set; }
+        public string OriginatorConversationId { get; set; }
+        public string ResponseDescription { get; set; }
+        public string MerchantRequestID { get; set; }
+        public string CheckoutRequestID { get; set; }
+        public string ResponseCode { get; set; }
+        public string ResultDesc { get; set; }
+        public string ResultCode { get; set; }
+        public string CustomerMessage { get; set; }
+    }
+}
