+++ MpesaDaraja/MpesaDaraja.csproj
@@ -11,4 +11,10 @@
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="Security\certificate.txt">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 </Project>
